Fall back to mysql-client-version in installation.
This is required because the server-version might not be available
without a connection. So, before the user has entered the mysql-
connection-settings, we might be unable to get the server-version.
This solution is not optimal because client- and server-version might be
different so that the displayed version changes, which would be
confusing. So, ideally, we should restructure the installation so that
the mysql-connection-settings are typed in first and afterwards the
mysql version is checked.