Consolidated db-connect code for install, front etc. (fixed #76)
This fixed the umlaut problem during installation, because we
used a simplified way to connect previously, which lacked a few
things that were done in BS itself. Now that we use the same code,
it is no problem anymore.
Important: charset for HTML and MySQL is now UTF8 by default in the
config/userdef.php. We don't use mysql_get_server_version() anymore
because this causes the problem that mysql_error() is empty for
errors during mysql_connect() anymore. And since it was only
necessary for mysql < 4.1, which shouldn't be in use anymore, we
decided that it's more important to have the error message.